All values moved over
